refactor(core): clean up utils folder
- Moved treasury code into the main utils folder
- Renamed the general utils file to datum.ts so it makes more sense
- Fixed all the imports to point to the new files
