+++ .gitignore
@@ -28,5 +28,7 @@ out/
 tmp/
 .env
 
+PATTERNS.md
+
 # SDK Tarballs (generated by pnpm pack / repack)
 *.tgz
