+++ docs/dcu-kit-design-specs/dcu-kit.pdf
@@ -41096,8 +41096,8 @@ endobj
 3543 0 obj
 <<
   /Creator (Typst 0.14.2)
-  /ModDate (D:20260222145536Z)
-  /CreationDate (D:20260222145536Z)
+  /ModDate (D:20260222205522Z)
+  /CreationDate (D:20260222205522Z)
 >>
 endobj
 
@@ -41108,7 +41108,7 @@ endobj
   /Subtype /XML
 >>
 stream
-<?xpacket begin="﻿" id="W5M0MpCehiHzreSzNTczkc9d"?><x:xmpmeta xmlns:x="adobe:ns:meta/" x:xmptk="xmp-writer"><rdf:RDF xmlns:rdf="http://www.w3.org/1999/02/22-rdf-syntax-ns#"><rdf:Description rdf:about="" xmlns:dc="http://purl.org/dc/elements/1.1/"  xmlns:xmp="http://ns.adobe.com/xap/1.0/"  xmlns:xmpMM="http://ns.adobe.com/xap/1.0/mm/"  xmlns:xmpTPg="http://ns.adobe.com/xap/1.0/t/pg/"  xmlns:pdf="http://ns.adobe.com/pdf/1.3/" ><xmp:CreatorTool>Typst 0.14.2</xmp:CreatorTool><dc:language><rdf:Bag><rdf:li>en</rdf:li></rdf:Bag></dc:language><xmp:ModifyDate>2026-02-22T14:55:36+00:00</xmp:ModifyDate><xmp:CreateDate>2026-02-22T14:55:36+00:00</xmp:CreateDate><xmpTPg:NPages>40</xmpTPg:NPages><dc:format>application/pdf</dc:format><xmpMM:InstanceID>LL7fxKkYrY26dkRI6Z/9RA==</xmpMM:InstanceID><xmpMM:DocumentID>LL7fxKkYrY26dkRI6Z/9RA==</xmpMM:DocumentID><xmpMM:RenditionClass>proof</xmpMM:RenditionClass><pdf:PDFVersion>1.7</pdf:PDFVersion></rdf:Description></rdf:RDF></x:xmpmeta><?xpacket end="r"?>
+<?xpacket begin="﻿" id="W5M0MpCehiHzreSzNTczkc9d"?><x:xmpmeta xmlns:x="adobe:ns:meta/" x:xmptk="xmp-writer"><rdf:RDF xmlns:rdf="http://www.w3.org/1999/02/22-rdf-syntax-ns#"><rdf:Description rdf:about="" xmlns:dc="http://purl.org/dc/elements/1.1/"  xmlns:xmp="http://ns.adobe.com/xap/1.0/"  xmlns:xmpMM="http://ns.adobe.com/xap/1.0/mm/"  xmlns:xmpTPg="http://ns.adobe.com/xap/1.0/t/pg/"  xmlns:pdf="http://ns.adobe.com/pdf/1.3/" ><xmp:CreatorTool>Typst 0.14.2</xmp:CreatorTool><dc:language><rdf:Bag><rdf:li>en</rdf:li></rdf:Bag></dc:language><xmp:ModifyDate>2026-02-22T20:55:22+00:00</xmp:ModifyDate><xmp:CreateDate>2026-02-22T20:55:22+00:00</xmp:CreateDate><xmpTPg:NPages>40</xmpTPg:NPages><dc:format>application/pdf</dc:format><xmpMM:InstanceID>UqltYBZF1iOzeRjacCqOJg==</xmpMM:InstanceID><xmpMM:DocumentID>UqltYBZF1iOzeRjacCqOJg==</xmpMM:DocumentID><xmpMM:RenditionClass>proof</xmpMM:RenditionClass><pdf:PDFVersion>1.7</pdf:PDFVersion></rdf:Description></rdf:RDF></x:xmpmeta><?xpacket end="r"?>
 endstream
 endobj
 
@@ -44688,7 +44688,7 @@ trailer
   /Size 3546
   /Root 3545 0 R
   /Info 3543 0 R
-  /ID [(LL7fxKkYrY26dkRI6Z/9RA==) (LL7fxKkYrY26dkRI6Z/9RA==)]
+  /ID [(UqltYBZF1iOzeRjacCqOJg==) (UqltYBZF1iOzeRjacCqOJg==)]
 >>
 startxref
 1623386
