+++ sdk/src/core/utils/index.ts
@@ -1,5 +1,6 @@
 export * from "./assets.js";
 export * from "./script.js";
 export * from "./tx.js";
-export * from "./utils.js";
+export * from "./datum.js";
+export * from "./treasury.js";
 export * from "./wallet.js";
